+++ frontend/src/locales/en-US.json
@@ -629,9 +629,15 @@
         "remove-user-permission": "Remove user permission?",
         "remove-sensitive-warning": "The will expose the column."
       },
+      "algorithms": {
+        "self": "Masking Algorithm",
+        "default": "Default",
+        "default-desc": "Use *** as data mask"
+      },
       "masking-level": {
         "self": "Masking Level",
         "selet-level": "Select level",
+        "masking_level_unspecified": "Follow global masking rules",
         "none": "None",
         "partial": "Partial",
         "full": "Full"
