+++ frontend/src/locales/zh-CN.json
@@ -627,8 +627,14 @@
         "remove-user-permission": "删除该用户权限？",
         "remove-sensitive-warning": "该数据会被设置为明文，不再脱敏。"
       },
+      "algorithms": {
+        "self": "脱敏算法",
+        "default": "默认",
+        "default-desc": "使用 *** 为数据遮挡"
+      },
       "masking-level": {
         "self": "脱敏程度",
+        "masking_level_unspecified": "遵循全局配置",
         "none": "明文",
         "partial": "半脱敏",
         "full": "全脱敏"
